refactor: move todo creation to a helper method
