+++ src/utils/route-helper.js
@@ -1,4 +1,6 @@
 const { v4: uuidv4 } = require('uuid');
+const { toBrazillianStandardTime } = require('./route-util')
+
 
 function createUser(body) {
   const user = body
@@ -8,4 +10,17 @@ function createUser(body) {
   return user
 }
 
-module.exports = createUser
+function createTodo(request) {
+  return {
+    title: request.body.title,
+    deadline: new Date(request.body.deadline),
+    id: uuidv4(),
+    done: false,
+    created_at: toBrazillianStandardTime(new Date())
+  }
+}
+
+module.exports = {
+  createUser,
+  createTodo
+}
